src/requirements.txt
# Core dependencies
fastapi>=0.68.0,<0.69.0
uvicorn[standard]>=0.15.0,<0.16.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.0.6
email-validator>=1.1.3,<2.0.0
python-dotenv>=0.19.0,<0.20.0
pydantic>=1.8.0,<2.0.0

# Database
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.1,<3.0.0
alembic>=1.7.3,<2.0.0

# GraphQL
strawberry-graphql>=0.92.0,<0.93.0

# Redis
redis>=3.5.3,<4.0.0

# DocuSign
docusign-esign>=3.10.0,<4.0.0

# Testing
pytest>=6.2.5,<7.0.0
pytest-asyncio>=0.15.1,<0.16.0
httpx>=0.19.0,<0.20.0
